Project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.11.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.4
celery==5.6.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
Django==6.0.1
fpdf==1.7.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
invoke==2.2.0
jiter==0.10.0
kombu==5.6.2
multidict==6.6.4
mysqlclient==2.2.7
openai==0.28.0
packaging==26.0
paramiko==4.0.0
pillow==11.2.1
prompt_toolkit==3.0.52
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.5
sshtunnel==0.4.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.3.2
yarl==1.20.1
reportlab==4.1.0
gunicorn==21.2.0
psycopg[binary]>=3.1.8
dj-database-url==2.1.0
whitenoise==6.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-ratelimit==4.1.0
drf-spectacular==0.28.0
